finished extracting logic into map.ts utility file
